Progress towards pure-rust prover, verifier
* GPIO_SENDRECV support
* GPIO_SHA support
* Support for ProverOpts
* circuit trace messages now format correctly, and can be enabled with RUST_LOG=trace

Now, many more of the tests pass with --features=pure-prove

+++ risc0/zkvm/sdk/rust/guest/src/env.rs
@@ -168,8 +168,8 @@ impl Env {
         // Write the full data out to the host
         unsafe {
             let desc = IoDescriptor {
-                size: len_bytes,
-                addr: slice.as_ptr() as usize,
+                size: len_bytes as u32,
+                addr: slice.as_ptr() as u32,
             };
             let ptr: *const IoDescriptor = &desc;
             memory_barrier(ptr);
